Lab 10 Answers (worked w/ Brandon Krakora)
1.) Discuss the differences among the results. Does the starting point make a difference?

The differences in my results is that DepthFirstIterator printed in reverse to BreadthFirstIterator.

I did not notice any differences; no difference between BreadthFirstIterator and ClosestFirstIterator.
In addition, RandomWalkIterator printed randomly.

BreadthFirstIterator and DepthFirstIterator printed the reversed results for the same starting vertice.

The starting point does make a difference.
2.) Give two or three examples of how/why a real-world map might require more than four colors.

A map in the real world probably might require more than four colors if there are multiple disconnected area/regions, like islands, more colors would be needed.
An additional example is if the map uses a color key to mark certain regions or areas.

Examples are of Countries now, not the states I created. It was easy to describe by looking on an actual map.
Example of Enclave: Vatican City is entirely surrounded by the territory of Italy. Different colors will be used for the two countries and neighbors of Italy.
Example of Exclave: Campione d'Italia, Province of Como, Italy is separated from the main area of Italy and is located in Switzerland's territory/area. Campoine will be painted the same color with Italy, however, Switzerland would be different, and neighbors of the two countries as well.
If several countries share or have the same corner, then more than four colors will be used.
Usually the real maps will use more than four colors to show different areas, and for a more appealing visual representation.

3.)RandomWalkIterator prints randomly on all the country names infinitely, no matter what the starting vertice is. The ClosestFirstIterator is a specific weighted path length starting a certain start vertex.The edge is considered in the actual calculations when added to get the path length.